Rename cli module to _detail.logging
This renaming is effectively two changes, based on code review feedback:

- The module is moved into a separate _detail package, to highlight that
  its contents are an implementation detail and not part of CBI's public
  interface.

- The module is renamed from "cli" to "logging", since the three classes
  it defines (Formatter, MetaWarning, WarningAggregator) are all related
  to logging.
