+++ codebasin/_detail/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (C) 2019-2025 Intel Corporation
+# SPDX-License-Identifier: BSD-3-Clause
+"""
+This package contains implementation details that are not part of the public
+interface of Code Base Investigator. These implementation details are not
+intended to be used by other scripts, and should not be relied upon.
+"""
+import codebasin._detail.logging
